Move test CSV files to parent package
The CSV files and the import function are now being used for test setup
stage and do not belong in the converters package.
