+++ test/integration/ggrc/test_csvs/all_snapshottable_objects.csv
