+++ test/integration/ggrc/test_csvs/assessment_with_templates.csv
