+++ test/integration/ggrc/test_csvs/facilities_mapping_game.csv
